A letter histogram is a graphical representation of number of occurances of different letters in a sentence. Your task is to draw a vertical letter histogram for a user given sentence. The histogram should not have any entry for the count of blank spaces.  Input Format  A sentence (String)  Constraints  The sentence contains letters, digits and blank spaces only.  Output Format  The vertical histogram. (Refer to sample for explanation)  Sample Input 0  this is a test Sample Output 0  a | * e | * h | * i | ** s | *** t | *** Explanation 0  The horizontal bars corresponding to each letter is depicted by the character '*'. Suppose the letter 'a' is present only one time in the input string, thus the bar corresponding to letter a has one '*'. Character 's' is present 3 times, hence the length of the bar is 3 '*' characters. There is no space between the axis characters (e.g. 'a'), the '|' character and the beginning of the bar. Note that the letters are alphabetically sorted along the vertical axis and all letters are in lower case.
A letter histogram is a graphical representation of number of occurances of different letters in a sentence. Your task is to draw a vertical letter histogram for a user given sentence. The histogram should not have any entry for the count of blank spaces.

Input Format

A sentence (String)

Constraints

The sentence contains letters, digits and blank spaces only.

Output Format

The vertical histogram. (Refer to sample for explanation)

Sample Input 0

this is a test
Sample Output 0

a | *
e | *
h | *
i | **
s | ***
t | ***
Explanation 0

The horizontal bars corresponding to each letter is depicted by the character '*'. Suppose the letter 'a' is present only one time in the input string, thus the bar corresponding to letter a has one '*'. Character 's' is present 3 times, hence the length of the bar is 3 '*' characters. There is no space between the axis characters (e.g. 'a'), the '|' character and the beginning of the bar. Note that the letters are alphabetically sorted along the vertical axis and all letters are in lower case.